+++ src/js/repeat.js
@@ -13,6 +13,7 @@
 
 var $ = require( 'jquery' );
 var config = require( 'enketo-config' );
+var disableFirstRepeatRemoval = config.repeatOrdinals === true;
 
 module.exports = {
     /**
@@ -29,15 +30,19 @@ module.exports = {
 
         $repeatInfos = this.form.view.$.find( '.or-repeat-info' );
         this.templates = {};
+        // Add repeat numbering elements
         $repeatInfos.siblings( '.or-repeat' )
             .prepend( '<span class="repeat-number"></span>' );
-        $repeatInfos.filter( '*:not([data-repeat-fixed]):not([data-repeat-count])' ).siblings( '.or-repeat' )
-            .append( '<div class="repeat-buttons"><button type="button" class="btn btn-default repeat"><i class="icon icon-plus"> </i></button>' +
-                '<button type="button" disabled class="btn btn-default remove"><i class="icon icon-minus"> </i></button></div>' );
-
+        // Add repeat buttons
+        $repeatInfos.filter( '*:not([data-repeat-fixed]):not([data-repeat-count])' )
+            .append( '<button type="button" class="btn btn-default add-repeat-btn"><i class="icon icon-plus"> </i></button>' )
+            .siblings( '.or-repeat' )
+            .append( '<div class="repeat-buttons">' +
+                '<button type="button" ' + ( disableFirstRepeatRemoval ? ' disabled ' : ' ' ) +
+                'class="btn btn-default remove"><i class="icon icon-minus"> </i></button></div>' );
         /**
          * The model also requires storing repeat templates for repeats that do not have a jr:template.
-         * Since the model has no knowledge of which node is a repeat, we do this here.
+         * Since the model has no knowledge of which node is a repeat, we direct this here.
          */
         this.form.model.extractFakeTemplates( $repeatInfos.map( function() {
             return this.dataset.name;
@@ -49,56 +54,97 @@ module.exports = {
          *
          * Widgets not yet initialized. Values not yet set.
          */
-        $repeatInfos.siblings( '.or-repeat' ).clone().reverse().each( function() {
-            var $templateEl = $( this );
+        $repeatInfos.siblings( '.or-repeat' ).reverse().each( function() {
+            var $templateEl = $( this ).remove();
             var xPath = $templateEl.attr( 'name' );
-            that.templates[ xPath ] = $templateEl;
+            that.templates[ xPath ] = $templateEl[ 0 ];
         } );
 
-        $repeatInfos.filter( '*:not([data-repeat-count])' ).each( function() {
+        $repeatInfos.filter( '*:not([data-repeat-count])' ).reverse()
+            .each( function() {
+                // don't do nested repeats here, they will be dealt with recursively
+                if ( !$( this ).closest( '.or-repeat' ).length ) {
+                    that.updateDefaultFirstRepeatInstance( null, this );
+                }
+            } )
             // If there is no repeat-count attribute, check how many repeat instances 
             // are in the model, and update view if necessary.
-            that.updateViewInstancesFromModel( $( this ) );
-
-        } );
+            .each( that.updateViewInstancesFromModel.bind( this ) );
 
         // delegated handlers (strictly speaking not required, but checked for doubling of events -> OK)
-        this.form.view.$.on( 'click', 'button.repeat:enabled', function() {
+        this.form.view.$.on( 'click', 'button.add-repeat-btn:enabled', function() {
             // Create a clone
-            that.clone( $( this ).closest( '.or-repeat' ).siblings( '.or-repeat-info' ).eq( 0 ) );
+            that.add( $( this ).closest( '.or-repeat-info' )[ 0 ] );
             // Prevent default
             return false;
         } );
         this.form.view.$.on( 'click', 'button.remove:enabled', function() {
             //remove clone
-            that.remove( $( this ).closest( '.or-repeat.clone' ) );
+            that.remove( $( this ).closest( '.or-repeat' ) );
             //prevent default
             return false;
         } );
 
         this.countUpdate();
     },
-    updateViewInstancesFromModel: function( $repeatInfo ) {
+    /**
+     * [updateViewInstancesFromModel description]
+     * @param  {[type]} idx           not used but part of jQuery.each
+     * @param   {Element} repeatInfo  repeatInfo element
+     * @return {[type]}            [description]
+     */
+    updateViewInstancesFromModel: function( idx, repeatInfo ) {
         var that = this;
-        var repeatPath = $repeatInfo.data( 'name' );
+        var $repeatInfo = $( repeatInfo );
+        var repeatPath = repeatInfo.dataset.name;
         // All we need is to find out in which series we are.
-        var repeatSeriesIndex = this.form.view.$.find( '.or-repeat-info[data-name="' + repeatPath + '"]' ).index( $repeatInfo );
+        var repeatSeriesIndex = this.form.view.$.find( '.or-repeat-info[data-name="' + repeatPath + '"]' ).index( repeatInfo );
         var repInModelSeries = this.form.model.getRepeatSeries( repeatPath, repeatSeriesIndex );
         var repInViewSeries = $repeatInfo.siblings( '.or-repeat' );
         // First rep is already included (by XSLT transformation)
         if ( repInModelSeries.length > repInViewSeries.length ) {
-            this.clone( $repeatInfo, repInModelSeries.length - repInViewSeries.length );
+            this.add( repeatInfo, repInModelSeries.length - repInViewSeries.length );
             // Now check the repeat counts of all the descendants of this repeat and its new siblings
             // Note: not tested with triple-nested repeats, but probably taking the better safe-than-sorry approach,
             // so should be okay except for performance.
             $repeatInfo.siblings( '.or-repeat' )
                 .find( '.or-repeat-info:not([data-repeat-count])' )
                 .each( function() {
-                    that.updateViewInstancesFromModel( $( this ) );
+                    that.updateViewInstancesFromModel( null, this );
                 } );
         }
+        return repInModelSeries.length;
     },
-    updateRepeatInstancesFromCount: function( $info ) {
+    /**
+     * [updateDefaultFirstRepeatInstance description]
+     * @param  {[type]} idx             not use but part of jQeury.each
+     * @param   {Element} repeatInfo    repeatInfo element
+     * @return {[type]}            [description]
+     */
+    updateDefaultFirstRepeatInstance: function( idx, repeatInfo ) {
+        var repeatSeriesIndex;
+        var repeatSeriesInModel;
+        var that = this;
+        var $repeatInfo = $( repeatInfo );
+        var repeatPath = repeatInfo.dataset.name;
+        if ( !that.form.model.data.instanceStr && !this.templates[ repeatPath ].classList.contains( 'or-appearance-minimal' ) ) {
+            repeatSeriesIndex = this.form.view.$.find( '.or-repeat-info[data-name="' + repeatPath + '"]' ).index( repeatInfo );
+            repeatSeriesInModel = this.form.model.getRepeatSeries( repeatPath, repeatSeriesIndex );
+            if ( repeatSeriesInModel.length === 0 ) {
+                this.add( repeatInfo );
+            }
+            $repeatInfo.siblings( '.or-repeat' )
+                .find( '.or-repeat-info:not([data-repeat-count])' )
+                .each( that.updateDefaultFirstRepeatInstance.bind( that ) );
+        }
+    },
+    /**
+     * [updateRepeatInstancesFromCount description]
+     * @param  {[type]} idx             not use but part of jQeury.each
+     * @param   {Element} repeatInfo repeatInfo element
+     * @return {[type]}            [description]
+     */
+    updateRepeatInstancesFromCount: function( idx, repeatInfo ) {
         var that = this;
         var $last;
         var repCountNodes;
@@ -107,7 +153,8 @@ module.exports = {
         var toCreate;
         var repPath;
         var repIndex;
-        var repCountPath = $info.data( 'repeatCount' ) || '';
+        var $repeatInfo = $( repeatInfo );
+        var repCountPath = repeatInfo.dataset.repeatCount || '';
 
         if ( !repCountPath ) {
             return;
@@ -123,8 +170,8 @@ module.exports = {
          * 3. select the correct node from the result array
          * 
          */
-        repPath = $info.data( 'name' );
-        repIndex = this.form.view.$.find( '.or-repeat-info[data-name="' + repPath + '"]' ).index( $info );
+        repPath = repeatInfo.dataset.name;
+        repIndex = this.form.view.$.find( '.or-repeat-info[data-name="' + repPath + '"]' ).index( repeatInfo );
         repCountNodes = this.form.model.evaluate( repCountPath, 'nodes', null, null, true );
 
         if ( repCountNodes.length && repCountNodes[ repIndex ] ) {
@@ -134,27 +181,24 @@ module.exports = {
         }
 
         numRepsInCount = isNaN( numRepsInCount ) ? 0 : numRepsInCount;
-        numRepsInView = $info.siblings( '.or-repeat[name="' + repPath + '"]' ).length;
+        numRepsInView = $repeatInfo.siblings( '.or-repeat[name="' + repPath + '"]' ).length;
         toCreate = numRepsInCount - numRepsInView;
 
         if ( toCreate > 0 ) {
-            that.clone( $info, toCreate );
+            that.add( repeatInfo, toCreate );
         } else if ( toCreate < 0 ) {
-            // temporary undocumented way to disable ability to remove first repeat
-            toCreate = Math.abs( toCreate ) >= numRepsInView ? -numRepsInView + ( config.repeatOrdinals === true ? 1 : 0 ) : toCreate;
+            toCreate = Math.abs( toCreate ) >= numRepsInView ? -numRepsInView + ( disableFirstRepeatRemoval ? 1 : 0 ) : toCreate;
             for ( ; toCreate < 0; toCreate++ ) {
-                $last = $info.siblings( '.or-repeat' ).last();
+                $last = $repeatInfo.siblings( '.or-repeat' ).last();
                 this.remove( $last, 0 );
             }
         }
         // Now check the repeat counts of all the descendants of this repeat and its new siblings, level-by-level.
         // TODO: this does not find .or-repeat > .or-repeat (= unusual syntax)
-        $info.siblings( '.or-repeat' )
+        $repeatInfo.siblings( '.or-repeat' )
             .children( '.or-group, .or-group-data' )
             .children( '.or-repeat-info[data-repeat-count]' )
-            .each( function() {
-                that.updateRepeatInstancesFromCount( $( this ) );
-            } );
+            .each( that.updateRepeatInstancesFromCount.bind( that ) );
     },
     /**
      * Checks whether repeat count value has been updated and updates repeat instances
@@ -165,61 +209,64 @@ module.exports = {
      */
     countUpdate: function( updated ) {
         var $repeatInfos;
-        var that = this;
 
         updated = updated || {};
         $repeatInfos = this.form.getRelatedNodes( 'data-repeat-count', '.or-repeat-info', updated );
 
-        $repeatInfos.each( function() {
-            that.updateRepeatInstancesFromCount( $( this ) );
-        } );
+        $repeatInfos.each( this.updateRepeatInstancesFromCount.bind( this ) );
     },
     /**s
      * clone a repeat group/node
-     * @param   {jQuery} $node node to clone
+     * @param   {Element} repeatInfo repeatInfo element
      * @param   {number=} count number of clones to create
      * @return  {boolean}       [description]
      */
-    clone: function( $info, count ) {
+    add: function( repeatInfo, count ) {
         var $repeats;
         var $clone;
-        var $repeatsToUpdate;
         var repeatIndex;
         var repeatSeriesIndex;
         var repeatPath;
         var i;
         var that = this;
+        var $repeatInfo = $( repeatInfo );
         var byCountUpdate = !!count;
         var modelRepeatSeries;
 
         count = count || 1;
 
-        if ( $info.length !== 1 ) {
+        if ( !repeatInfo ) {
             console.error( 'Nothing to clone' );
             return false;
         }
 
-        repeatPath = $info.data( 'name' );
-        $repeats = $info.siblings( '.or-repeat' );
+        repeatPath = repeatInfo.dataset.name;
+        $repeats = $repeatInfo.siblings( '.or-repeat' );
 
-        // Add clone class to allow some simplistic performance optimization elsewhere.
-        $clone = this.templates[ repeatPath ].clone().addClass( 'clone' );
+        $clone = $( this.templates[ repeatPath ] ).clone();
+
         // Determine the index of the repeat series.
-        repeatSeriesIndex = this.form.view.$.find( '.or-repeat-info[data-name="' + repeatPath + '"]' ).index( $info );
+        repeatSeriesIndex = this.form.view.$.find( '.or-repeat-info[data-name="' + repeatPath + '"]' ).index( repeatInfo );
 
         // Add required number of repeats
         for ( i = 0; i < count; i++ ) {
             // Fix names of radio button groups
             $clone.find( '.option-wrapper' ).each( this.fixRadioNames );
             // Insert the clone
-            $clone.insertBefore( $info );
+            $clone.insertBefore( repeatInfo );
+
+            if ( $clone.prev( '.or-repeat' ).length ) {
+                // Also add the clone class for all 2+ numbers as this is
+                // use for performance optimization in several places.
+                $clone.addClass( 'clone' );
+            }
             // Update the variable containing the view repeats in the current series.
             $repeats = $repeats.add( $clone );
             // TODO: not efficient, could be incremental
             modelRepeatSeries = this.form.model.getRepeatSeries( repeatPath, repeatSeriesIndex );
             // Create a repeat in the model if it doesn't already exist
             if ( $repeats.length > modelRepeatSeries.length ) {
-                this.form.model.cloneRepeat( repeatPath, repeatSeriesIndex );
+                this.form.model.addRepeat( repeatPath, repeatSeriesIndex );
             }
             // This is the index of the new repeat in relation to all other repeats of the same name,
             // even if they are in different series.
@@ -238,16 +285,18 @@ module.exports = {
                     repeatIndex: repeatIndex
                 } );
             }
+            // now create the first instance of any nested repeats if necessary
+            $clone.find( '.or-repeat-info:not([data-repeat-count])' ).each( this.updateDefaultFirstRepeatInstance.bind( this ) );
 
-            $clone = this.templates[ repeatPath ].clone().addClass( 'clone' );
+            $clone = $( this.templates[ repeatPath ] ).clone();
+
             repeatIndex++;
         }
 
-        $repeatsToUpdate = $repeats.add( $repeats.find( '.or-repeat' ) );
         // number the repeats
-        this.numberRepeats( $repeatsToUpdate );
+        this.numberRepeats( repeatInfo );
         // enable or disable + and - buttons
-        this.toggleButtons( $repeatsToUpdate );
+        this.toggleButtons( repeatInfo );
 
         return true;
     },
@@ -256,14 +305,14 @@ module.exports = {
         var $prev = $repeat.prev( '.or-repeat' );
         var repeatPath = $repeat.attr( 'name' );
         var repeatIndex = this.form.view.$.find( '.or-repeat[name="' + repeatPath + '"]' ).index( $repeat );
-        var $siblings = $repeat.siblings( '.or-repeat' );
+        var repeatInfo = $repeat.siblings( '.or-repeat-info' )[ 0 ];
 
         delay = typeof delay !== 'undefined' ? delay : 600;
 
         $repeat.hide( delay, function() {
             $repeat.remove();
-            that.numberRepeats( $siblings );
-            that.toggleButtons( $siblings );
+            that.numberRepeats( repeatInfo );
+            that.toggleButtons( repeatInfo );
             // Trigger the removerepeat on the previous repeat (always present)
             // so that removerepeat handlers know where the repeat was removed
             $prev.trigger( 'removerepeat' );
@@ -275,44 +324,20 @@ module.exports = {
         $( element ).find( 'input[type="radio"]' )
             .attr( 'name', Math.floor( ( Math.random() * 10000000 ) + 1 ) );
     },
-    toggleButtons: function( $repeats ) {
-        var $repeat;
-        var $repSiblingsAndSelf;
-
-        $repeats = ( !$repeats || $repeats.length === 0 ) ? $() : $repeats;
-
-        $repeats.each( function() {
-            $repeat = $( this );
-            $repSiblingsAndSelf = $repeat.siblings( '.or-repeat' ).addBack();
-            //first switch everything off and remove hover state
-            $repSiblingsAndSelf.children( '.repeat-buttons' ).find( 'button.repeat, button.remove' ).prop( 'disabled', true );
-
-            //then enable the appropriate ones
-            $repSiblingsAndSelf.last().children( '.repeat-buttons' ).find( 'button.repeat' ).prop( 'disabled', false );
-            $repSiblingsAndSelf.children( '.repeat-buttons' ).find( 'button.remove' ).not( ':first' ).prop( 'disabled', false );
-        } );
+    toggleButtons: function( repeatInfo ) {
+        $( repeatInfo )
+            .siblings( '.or-repeat' )
+            .children( '.repeat-buttons' )
+            .find( 'button.remove' )
+            .prop( 'disabled', false )
+            .first()
+            .prop( 'disabled', disableFirstRepeatRemoval );
     },
-    numberRepeats: function( $repeats ) {
-        $repeats.each( function() {
-            var $repSiblings;
-            var qtyRepeats;
-            var i;
-            var $repeat = $( this );
-            // if it is the first-of-type (not that ':first-of-type' does not have cross-browser support)
-            if ( $repeat.prev( '.or-repeat' ).length === 0 ) {
-                $repSiblings = $( this ).siblings( '.or-repeat' );
-                qtyRepeats = $repSiblings.length + 1;
-                if ( qtyRepeats > 1 ) {
-                    $repeat.find( '.repeat-number' ).text( '1' );
-                    i = 2;
-                    $repSiblings.each( function() {
-                        $( this ).find( '.repeat-number' ).eq( 0 ).text( i );
-                        i++;
-                    } );
-                } else {
-                    $repeat.find( '.repeat-number' ).eq( 0 ).empty();
-                }
-            }
-        } );
+    numberRepeats: function( repeatInfo ) {
+        $( repeatInfo )
+            .siblings( '.or-repeat' )
+            .each( function( idx, repeat ) {
+                $( repeat ).children( '.repeat-number' ).text( idx + 1 );
+            } );
     }
 };