+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ##### Added
 - Make all label groups collapsible.
 - Let appearance "compact" on a group collapse this group by default.
+- Make first repeat removable with button if repeat-count is not used.
+- Let appearance "minimal" on a repeat prevent automatic creation of the first repeat instance.
 
 [4.39.5] - 2017-10-09
 ----------------------